添加管理员管理功能，包括检查、删除和注册用户账号的实现，更新 Dri_NVS 模块以修复读取字符串的参数类型

+++ main/Driver/Dri_NVS.h
@@ -23,7 +23,7 @@ esp_err_t Dri_NVS_WriteStr(uint8_t *key, uint8_t *value);
  * @param value 键值
  * @param value_len 键值长度
  */
-esp_err_t Dri_NVS_ReadStr(uint8_t *key, uint8_t *value, size_t *value_len);
+esp_err_t Dri_NVS_ReadStr(uint8_t *key, uint8_t *value, uint8_t *value_len);
 
 /**
  * @brief 删除 NVS 中的指定键
